random: separate random.c and init.c
This requires less code to be compiled when recompiling init.c on each
build with a new RANDOM_SEED_DEFAULT value.
